requirements.txt
Flask>=2.0.0
opencv-python-headless>=4.5.0
pyzbar>=0.1.9
numpy>=1.21.0
Pillow>=9.0.0
python-dateutil>=2.8.0
pymongo>=4.0.0
gunicorn>=20.1.0
